+++ src/List.c
@@ -8,7 +8,7 @@
 
 #include <stdlib.h>
 
-List_t* list_new(size_t elem_size, freer_cb freer) {
+List_t* list_new(unsigned long elem_size, freer_cb freer) {
     List_t* out = malloc(sizeof(List_t));
     out->elem_size = elem_size;
